requirements.txt
flask
waitress
python-dotenv
requests
python-docx
pywin32